requirements.txt
pytest
httpx
allure-pytest
jsonschema
pydantic 
pydantic_settings
python-dotenv
faker